Refactor: move getAccount to accounts package (from openrtb2) (#1483)
