+++ endpoints/openrtb2/auction.go
@@ -22,6 +22,7 @@ import (
 	"github.com/mxmCherry/openrtb"
 	"github.com/mxmCherry/openrtb/native"
 	nativeRequests "github.com/mxmCherry/openrtb/native/request"
+	accountService "github.com/prebid/prebid-server/account"
 	"github.com/prebid/prebid-server/analytics"
 	"github.com/prebid/prebid-server/config"
 	"github.com/prebid/prebid-server/errortypes"
@@ -156,7 +157,7 @@ func (deps *endpointDeps) Auction(w http.ResponseWriter, r *http.Request, _ http
 	}
 
 	// Look up account now that we have resolved the pubID value
-	account, acctIDErrs := deps.getAccount(ctx, labels.PubID)
+	account, acctIDErrs := accountService.GetAccount(ctx, deps.cfg, deps.accounts, labels.PubID)
 	if len(acctIDErrs) > 0 {
 		errL = append(errL, acctIDErrs...)
 		writeError(errL, w, &labels)
@@ -1316,57 +1317,4 @@ func getAccountID(pub *openrtb.Publisher) string {
 		}
 	}
 	return pbsmetrics.PublisherUnknown
-}
-
-func (deps *endpointDeps) getAccount(ctx context.Context, pubID string) (account *config.Account, errs []error) {
-	// Check BlacklistedAcctMap until we have deprecated it
-	if _, found := deps.cfg.BlacklistedAcctMap[pubID]; found {
-		return nil, []error{&errortypes.BlacklistedAcct{
-			Message: fmt.Sprintf("Prebid-server has disabled Account ID: %s, please reach out to the prebid server host.", pubID),
-		}}
-	}
-	if deps.cfg.AccountRequired && pubID == pbsmetrics.PublisherUnknown {
-		return nil, []error{&errortypes.AcctRequired{
-			Message: fmt.Sprintf("Prebid-server has been configured to discard requests without a valid Account ID. Please reach out to the prebid server host."),
-		}}
-	}
-	if accountJSON, accErrs := deps.accounts.FetchAccount(ctx, pubID); len(accErrs) > 0 || accountJSON == nil {
-		// pubID does not reference a valid account
-		if len(accErrs) > 0 {
-			errs = append(errs, errs...)
-		}
-		if deps.cfg.AccountRequired && deps.cfg.AccountDefaults.Disabled {
-			errs = append(errs, &errortypes.AcctRequired{
-				Message: fmt.Sprintf("Prebid-server has been configured to discard requests without a valid Account ID. Please reach out to the prebid server host."),
-			})
-			return nil, errs
-		}
-		// Make a copy of AccountDefaults instead of taking a reference,
-		// to preserve original pubID in case is needed to check NonStandardPublisherMap
-		pubAccount := deps.cfg.AccountDefaults
-		pubAccount.ID = pubID
-		account = &pubAccount
-	} else {
-		// pubID resolved to a valid account, merge with AccountDefaults for a complete config
-		account = &config.Account{}
-		completeJSON, err := jsonpatch.MergePatch(deps.cfg.AccountDefaultsJSON(), accountJSON)
-		if err == nil {
-			err = json.Unmarshal(completeJSON, account)
-		}
-		if err != nil {
-			errs = append(errs, err)
-			return nil, errs
-		}
-		// Fill in ID if needed, so it can be left out of account definition
-		if len(account.ID) == 0 {
-			account.ID = pubID
-		}
-	}
-	if account.Disabled {
-		errs = append(errs, &errortypes.BlacklistedAcct{
-			Message: fmt.Sprintf("Prebid-server has disabled Account ID: %s, please reach out to the prebid server host.", pubID),
-		})
-		return nil, errs
-	}
-	return
 }
