+++ endpoints/openrtb2/auction_test.go
@@ -1984,8 +1984,7 @@ func (cf mockStoredReqFetcher) FetchRequests(ctx context.Context, requestIDs []s
 }
 
 var mockAccountData = map[string]json.RawMessage{
-	"valid_acct":    json.RawMessage(`{"disabled":false}`),
-	"disabled_acct": json.RawMessage(`{"disabled":true}`),
+	"valid_acct": json.RawMessage(`{"disabled":false}`),
 }
 
 type mockAccountFetcher struct {
@@ -2057,71 +2056,4 @@ type hardcodedResponseIPValidator struct {
 
 func (v hardcodedResponseIPValidator) IsValid(net.IP, iputil.IPVersion) bool {
 	return v.response
-}
-
-func TestGetAccount(t *testing.T) {
-	unknown := pbsmetrics.PublisherUnknown
-	testCases := []struct {
-		accountID string
-		// account_required
-		required bool
-		// account_defaults.disabled
-		disabled bool
-		// expected error, or nil if account should be found
-		err error
-	}{
-		// Blacklisted account is always rejected even in permissive setup
-		{accountID: "bad_acct", required: false, disabled: false, err: &errortypes.BlacklistedAcct{}},
-
-		// empty pubID
-		{accountID: unknown, required: false, disabled: false, err: nil},
-		{accountID: unknown, required: true, disabled: false, err: &errortypes.AcctRequired{}},
-		{accountID: unknown, required: false, disabled: true, err: &errortypes.BlacklistedAcct{}},
-		{accountID: unknown, required: true, disabled: true, err: &errortypes.AcctRequired{}},
-
-		// pubID given but is not a valid host account (does not exist)
-		{accountID: "doesnt_exist_acct", required: false, disabled: false, err: nil},
-		{accountID: "doesnt_exist_acct", required: true, disabled: false, err: nil},
-		{accountID: "doesnt_exist_acct", required: false, disabled: true, err: &errortypes.BlacklistedAcct{}},
-		{accountID: "doesnt_exist_acct", required: true, disabled: true, err: &errortypes.AcctRequired{}},
-
-		// pubID given and matches a valid host account with Disabled: false
-		{accountID: "valid_acct", required: false, disabled: false, err: nil},
-		{accountID: "valid_acct", required: true, disabled: false, err: nil},
-		{accountID: "valid_acct", required: false, disabled: true, err: nil},
-		{accountID: "valid_acct", required: true, disabled: true, err: nil},
-
-		// pubID given and matches a host account explicitly disabled (Disabled: true on account json)
-		{accountID: "disabled_acct", required: false, disabled: false, err: &errortypes.BlacklistedAcct{}},
-		{accountID: "disabled_acct", required: true, disabled: false, err: &errortypes.BlacklistedAcct{}},
-		{accountID: "disabled_acct", required: false, disabled: true, err: &errortypes.BlacklistedAcct{}},
-		{accountID: "disabled_acct", required: true, disabled: true, err: &errortypes.BlacklistedAcct{}},
-	}
-
-	for _, test := range testCases {
-		description := fmt.Sprintf(`ID=%s/required=%t/disabled=%t`, test.accountID, test.required, test.disabled)
-		t.Run(description, func(t *testing.T) {
-			deps := &endpointDeps{
-				cfg: &config.Configuration{
-					BlacklistedAcctMap: map[string]bool{"bad_acct": true},
-					AccountRequired:    test.required,
-					AccountDefaults:    config.Account{Disabled: test.disabled},
-				},
-				accounts: &mockAccountFetcher{},
-			}
-			assert.NoError(t, deps.cfg.MarshalAccountDefaults())
-
-			account, errors := deps.getAccount(context.Background(), test.accountID)
-
-			if test.err == nil {
-				assert.Empty(t, errors)
-				assert.Equal(t, test.accountID, account.ID, "account.ID must match requested ID")
-				assert.Equal(t, false, account.Disabled, "returned account must not be disabled")
-			} else {
-				assert.NotEmpty(t, errors, "expected errors but got success")
-				assert.Nil(t, account, "return account must be nil on error")
-				assert.IsType(t, test.err, errors[0], "error is of unexpected type")
-			}
-		})
-	}
 }
